main movie list shows party per movie and ratings

+++ app/src/main/java/student/rmit/edu/au/s3110401mad_assignment/model/Movie.java
@@ -18,7 +18,7 @@ public interface Movie {
 
     int getPoster();
 
-    float getRating();
+    int getRating();
 
-    void setRating(float rating);
+    void setRating(int rating);
 }
